Crate erudite

Structs

CommandConfig

Runner

Runner builder for a suite of tests

SimpleOutput

Data which can be returned from a command

TestCase

A test case which has an input and expected output

Enums

Bytes

Represents some data that may either be a string or a series of bytes. The recommended method for constructing this type is to use From::from which will automatically choose the appropriate variant for the data.

RunOutput

Represents the output of a Runner’s execution.

TestFailReason

The reason that a test may fail

TestOutput

A Result-like enum that represents a passed or failed test