erudite

Module runner

Definitions and functions related to the execution of test suites

Structs

CompileResult

The result from running the compile step

CompiledTestRunner

A test runner that has been compiled already. The tests are now able to be run.

TestHandle

A handle to a set of running tests. Tests can either be waited on one-at-a-time (using TestHandle::wait_next), or in bulk (using TestHandle::wait_all).

TestResult

The result from running the test. This also contains the data associated with the test and the index of the test when added to the TestContext.

TestRunner

A suite of tests that are about to be run. This can be created from the TestContext::test_runner function and will inherit the configuration from the context.

Enums

CompileResultState

The state of the result of the compilation step

TestResultState

The state of a test result