erudite

Struct Runner

pub struct Runner { /* private fields */ }

Runner builder for a suite of tests

let run_rules = Rules::new()
    .add_read_only("/usr")
    .add_read_only("/etc")
    .add_read_only("/bin");

let output = Runner::new()
    .compile_command(["ghc", "solution.hs"])
    .run_command(["./solution"])
    .test("hello", "olleh")
    .copy_file("solution.hs", "solutions/solution.hs")
    .run_rules(run_rules)
    .run()
    .await?;

Implementations

impl Runner

pub fn new() -> Self

Create a new runner with default values.

let mut runner = Runner::new();

pub fn compile_command( &mut self, command: impl IntoIterator<Item = impl Into<String>>, ) -> &mut Self

Set the command used to compile a solution for this runner. If the compile command is not specified, the compilation step will be skipped when calling Runner::run

This command is run in a custom directory that may be specified via this builder. See Runner::run_directory for more details.

let mut runner = Runner::new();
runner.compile_command(["javac", "Solution.java"]);

pub fn run_command( &mut self, command: impl IntoIterator<Item = impl Into<String>>, ) -> &mut Self

Set the command used to run a solution. If the run command is not specified, an error will be returned upon a call to Runner::run

This command is run in a custom directory that may be specified via this builder. See Runner::run_directory for more details.

let mut runner = Runner::new();
runner.run_command(["java", "Solution"]);

pub fn test( &mut self, input: impl Into<String>, output: impl Into<String>, ) -> &mut Self

Add a single test with expected input and expected output

If no tests are specified, an error is returned from Runner::run.

let mut runner = Runner::new();
runner.test("hello", "olleh")
      .test("hello world", "dlrow olleh");

pub fn tests(&mut self, cases: impl IntoIterator<Item = TestCase>) -> &mut Self

Add multiple tests to the list of test cases for this runner

If no tests are specified, an error is returned from Runner::run.

let mut runner = Runner::new();
runner.tests([
    TestCase::new("hello", "olleh"),
    TestCase::new("hello world", "dlrow olleh")
]);

pub fn trim_output(&mut self, trim_output: bool) -> &mut Self

Determines whether the output should be trimmed before comparison with the expected output (via slice::trim_ascii)

Default: true

pub fn timeout(&mut self, timeout: Duration) -> &mut Self

Set the maximum amount of time that any test is allowed to run

Note: this is for each test individually, not the whole suite

Default: 1 minute

pub fn run_directory(&mut self, path: impl Into<PathBuf>) -> &mut Self

Set the directory in which the tests will be run. If not specified, a directory will be created in std::env::temp_dir.

The program will be granted read/write/execute permissions in this directory

pub fn create_file( &mut self, path: impl Into<PathBuf>, content: impl AsRef<[u8]>, ) -> &mut Self

Designate a file to be created in the run directory with specific content

The path is based in the directory created in Runner::run_directory.

let mut runner = Runner::new();
runner.create_file("solution.py", r#"print("hello from python")"#);

pub fn copy_file( &mut self, dest: impl Into<PathBuf>, source: impl Into<PathBuf>, ) -> &mut Self

Designate a file to be copied into the run directory from a given location

The destination path is based in the directory created in Runner::run_directory. The source path is based on the current directory of the caller of this library.

let mut runner = Runner::new();
runner.copy_file("solution.hs", "example/solution.hs");

pub fn run_rules(&mut self, rules: Rules) -> &mut Self

Set the Rules for the run command.

If this is not specified, but compile_rules is set, then it will use those rules.
If neither is specified no rules will be placed on the run command.

Note: the rules provided will automatically have the run_directory added with read/write permissions.

pub fn compile_rules(&mut self, rules: Rules) -> &mut Self

Set the Rules for the run command.

If this is specified, but run_rules is not set, then run_rules will use these rules.
If this is not specified, then the compile command will be run with no rules set.

Note: the rules provided will automatically have the run_directory added with read/write permissions.

pub fn max_memory(&mut self, max_memory: CommandConfig<MemorySize>) -> &mut Self

Limit the maximum amount of memory that the commands can use

pub fn max_file_size( &mut self, max_file_size: CommandConfig<MemorySize>, ) -> &mut Self

Limit the maximum file size that the command can produce

pub async fn run(&mut self) -> Result<RunOutput>

Build and run all tests associated with this runner

Will error if the runner is not in a valid state or an error occurs while setting up or cleaning up the tests.

Any output related to the execution of the tests will be returned throught the RunOutput