erudite

Struct TestContextBuilder

pub struct TestContextBuilder<G, T, Tests = MissingTests, RunCmd = MissingRunCmd> { /* private fields */ }

A builder that will crate a TestContext. This builder uses a type-state pattern to ensure that the require fields are all present.

The fields timeout, rules, max_memory, max_file_size, and max_threads can be set for runtime, compile-time, or both. These methods have a pattern of run_<field> for runtime only, compile_<field> for compile-time only, or <field> for both runtime and compile-time.

Usage

let context = TestContext::builder()
    .compile_command(["rustc", "-o", "main", "main.rs"])
    .run_command(["./main"])   // Run command is always required
    .test("group", "hello", "olleh", 1) // At least one test is required
    .test("group", "world", "dlrow", 2)
    .test("group", "rust", "tsur", 3)
    .test("group", "tacocat", "tacocat", 4)
    .trim_output(true)
    .file(FileContent::string("// some rust code"), "main.rs")
    .timeout(Duration::from_secs(5))        // or `run_timeout`/`compile_timeout`
    .rules(rules)                           // or `run_rules`/`compile_rules`
    .max_memory(MemorySize::from_gib(1))    // or `run_max_memory`/`compile_max_memory`
    .max_file_size(MemorySize::from_gib(1)) // or `run_max_file_size`/`compile_max_file_size`
    .max_threads(2)                         // or `run_max_threads`/`compile_max_threads`
    .build();

Implementations

impl<G, T, Tests, RunCmd> TestContextBuilder<G, T, Tests, RunCmd>

pub fn trim_output(self, trim_output: bool) -> Self

Set whether the tests should trim the output of the program before testing (using str::trim)

Default = false

let ctx = TestContext::builder()
    .compile_command(["gcc", "-o", "solution", "solution.c"])
    .run_command(["solution.c"])
    .test("group", "hello world", "dlrow olleh", true)
    .trim_output(true)
    .build();

pub fn compile_command( self, compile_command: impl IntoIterator<Item = impl Into<String>>, ) -> Self

Set the command used for compilation of the test. If this function is not called, then no compilation step will be performed, even if the other compile settings are used.

The command must have at least one item to be valid. If it does not, the error will be propogated when calling TestRunner::compile

let ctx = TestContext::builder()
    .compile_command(["gcc", "-o", "solution", "solution.c"])
    .run_command(["solution.c"])
    .test("group", "hello world", "dlrow olleh", true)
    .build();

pub fn file( self, source: impl Into<FileContent>, destination: impl AsRef<Path>, ) -> Self

Add a file to be inserted into the runtime environment of the test. This is added before compilation, so it works well for libraries or input/output manipulation.

If source is a path, the file will only be copied when a test is compiled, meaning that if the file changes or is removed, then the output may differ between two instances of the test runner from a single context.

If the intended behaviour is to read the file now, consider reading the file directly and placing it into FileContent::bytes.

destination is path relative to the directory used for the test environment. If destination an absolute path, then it will be made relative to the test environment, i.e., /foo/bar -> <test-env>/foo/bar.

let ctx = TestContext::builder()
    .run_command(["node", "solution.js"])
    .test("group", "hello world", "dlrow olleh", true)
    .file(FileContent::string("// some javascript code"), "solution.js")
    .file(FileContent::path("/foo/bar"), "some_other_file")
    .build();

pub fn files( self, files: impl IntoIterator<Item = impl Into<FileConfig>>, ) -> Self

Add multiple files to be inserted into the runtime environment of the test. These files are added before compilation, so it works well for libraries or input/output manipulation.

If source is a path, the file will only be copied when a test is compiled, meaning that if the file changes or is removed, then the output may differ between two instances of the test runner from a single context.

If the intended behaviour is to read the file now, consider reading the file directly and placing it into FileContent::bytes.

destination is path relative to the directory used for the test environment. If destination an absolute path, then it will be made relative to the test environment, i.e., /foo/bar -> <test-env>/foo/bar.

let ctx = TestContext::builder()
    .run_command(["node", "solution.js"])
    .test("group", "hello world", "dlrow olleh", true)
    .files([
        FileConfig::new(FileContent::string("// some javascript code"), "solution.js"),
        FileConfig::new(FileContent::path("/foo/bar"), "some_other_file")
    ])
    .build();

pub fn run_rules(self, rules: Rules) -> Self

Set the rules for execution when running the test

pub fn compile_rules(self, rules: Rules) -> Self

Set the rules for execution when compiling the test

pub fn rules(self, rules: Rules) -> Self

Set the rules for execution of both running and compiling the test

pub fn run_max_memory(self, max_memory: MemorySize) -> Self

Set the maximum memory usage when running the test

pub fn compile_max_memory(self, max_memory: MemorySize) -> Self

Set the maximum memory usage when compiling the test

pub fn max_memory(self, max_memory: MemorySize) -> Self

Set the maximum memory usage of both running and compiling the test

pub fn run_max_file_size(self, max_file_size: MemorySize) -> Self

Set the maximum file size when running the test

pub fn compile_max_file_size(self, max_file_size: MemorySize) -> Self

Set the maximum file size when compiling the test

pub fn max_file_size(self, max_file_size: MemorySize) -> Self

Set the maximum file size of both running and compiling the test

pub fn run_max_threads(self, max_threads: u64) -> Self

Set the max thread count when running the test

pub fn compile_max_threads(self, max_threads: u64) -> Self

Set the max thread count when compiling the test

pub fn max_threads(self, max_threads: u64) -> Self

Set the max thread count of both running and compiling the test

pub fn run_timeout(self, timeout: Duration) -> Self

Set the timeout when running the test

pub fn compile_timeout(self, timeout: Duration) -> Self

Set the timeout when compiling the test

pub fn timeout(self, timeout: Duration) -> Self

Set the timeout of both running and compiling the test

impl<G, T, Tests> TestContextBuilder<G, T, Tests, MissingRunCmd>

pub fn run_command( self, command: impl IntoIterator<Item = impl Into<String>>, ) -> TestContextBuilder<G, T, Tests, SetRunCmd>

Set the command to run the tests

The command must have at least one item to be valid. If it does not, the error will be propogated when calling TestHandle::wait_next

let ctx = TestContext::builder()
    .run_command(["node", "solution.js"])
    .test("group", "hello world", "dlrow olleh", true)
    .build();

impl<G, T, Tests, RunCmd> TestContextBuilder<G, T, Tests, RunCmd>
where T: Clone, G: Hash + Eq,

pub fn test( self, group: G, input: impl Into<String>, output: impl Into<ExpectedOutput>, data: T, ) -> TestContextBuilder<G, T, SetTests, RunCmd>

Add a single test to this context in a group

The data argument is for any data that you wish to associate with this test (on top of its index, which is assocated by default). This is useful for adding additional information such as visiblity to the test.

The data is accessible by TestRunner::filter_tests and TestResult::data.

The group argument is used to associate tests with eachother. When creating a test runner, one uses the group to select just the tests that they wish to run. Group keys must implement Hash and Eq, but other than that, there is no restriction. This function requires an owned G for each test case, but to reduce copies, TestContextBuilder::tests may be preferred.

There is a special case if G is (): a TestContext::default_test_runner becomes available, which requires no argment and does not return an option. This is ideal if only one test group is necessary.

let ctx = TestContext::builder()
    .run_command(["echo", "hi"])
    .test("group", "hello world", "dlrow olleh", true)
    .build();

pub fn tests( self, group: G, tests: impl IntoIterator<Item = impl Into<TestCase<T>>>, ) -> TestContextBuilder<G, T, SetTests, RunCmd>

Add a series of tests to this context

The group argument is used to associate tests with eachother. When creating a test runner, one uses the group to select just the tests that they wish to run. Group keys must implement Hash and Eq, but other than that, there is no restriction. This function may be preferable over TestContextBuilder::test as there is only need for one allocated group key.

There is a special case if G is (): a TestContext::default_test_runner becomes available, which requires no argment and does not return an option. This is ideal if only one test group is necessary.

See Self::test for additional information.

let ctx = TestContext::builder()
    .run_command(["echo", "hi"])
    .tests("group", [
        ("hello world", "dlrow olleh", true),
        ("good morning", "gninrom doog", true),
    ])
    .build();

pub fn test_groups( self, test_groups: impl IntoIterator<Item = (G, impl IntoIterator<Item = impl Into<TestCase<T>>>)>, ) -> TestContextBuilder<G, T, SetTests, RunCmd>

Add several test groups to this context

The group argument is used to associate tests with eachother. When creating a test runner, one uses the group to select just the tests that they wish to run. Group keys must implement Hash and Eq, but other than that, there is no restriction. This function may be preferable over TestContextBuilder::test as there is only need for one allocated group key.

There is a special case if G is (): a TestContext::default_test_runner becomes available, which requires no argment and does not return an option. This is ideal if only one test group is necessary.

NOTE: This gets very syntactically noisy when writing out by hand. This function is very much intended for use with iterators, rather than manually written. If you’re manully writing cases, see TestContextBuilder::tests.

let ctx = TestContext::builder()
    .run_command(["echo", "hi"])
    .test_groups([
        ("group1", [("hello world", "dlrow olleh", true), ("good morning", "gninrom doog", true)]),
        ("group2", [("52", "even", true), ("27", "odd", true)])
    ])
    .build();

impl<G, T> TestContextBuilder<G, T, SetTests, SetRunCmd>
where G: Hash + Eq,

pub fn build(self) -> TestContext<G, T>

Finish building this test context